Atualização da Venda(Correção de Bugs) E funcionalidade de Consulta de Vendas

+++ mavenproject1/pi-2019/src/main/java/Modal/Livro.java
@@ -109,7 +109,7 @@ public class Livro {
         this.Quantidade = Quantidade;
     }
 
-    public double formatToFloat(String preco) {
+    public float formatToFloat(String preco) {
         float precoNovo;
         String precoFormatado;
         precoFormatado = preco.replaceAll("\\.", "").replaceAll("\\,", ".").replaceAll("R", "").replaceAll("\\$", "");
